Renamed GetToolTranslationsUseCase to GetToolTranslationsFilesUseCase
