+++ godtools/App/Share/Domain/UseCases/GetToolTranslationsFilesUseCase/ToolTranslationsDomainModel.swift
